requirements.txt
asgiref==3.7.2
Django~=4.2.4
sqlparse==0.4.4
typing_extensions==4.7.1
celery~=5.3.1
redis
utils~=1.0.1
requests~=2.31.0
psycopg2
pillow
ipython
django4-background-tasks
python-dotenv
django_crontab
allauth
guardian